refacto: predictor name + tests
